requirements.txt
functions-framework==3.*
requests
python-dotenv
google-cloud-secret-manager